fix: corrected logic in shop menu component
